extract ExpirationEntry
